+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Pre-release changes - please put everything in the appropriate category below
 
 ### Enhancements
+- GP2-3336 - Product classification expander on dashboard overview page plus content changes
+- GP2-3395 - Content changes and move hs code on product finder
 - GP2-3338 - Dashboard top 3 export plans
 - GP2-2574 - WTE add/remove markets to or from shortlist
 - GP2-3344 - Verify account with token